Move SQLOrderByExpr and Join* to query.rs (2/5)
